se realiza logica para crud y se conecta botones de editar agregar y eliminar

+++ src/app/components/personas/personas.component.ts
@@ -3,6 +3,7 @@ import { Crud } from '../../interfaces/crud';
 import { RouterModule } from '@angular/router';
 import { PersonasService } from '../../services/personas/personas.service';
 import { ProgressBarComponent } from "../../shared/progress-bar/progress-bar.component";
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-personas',
@@ -20,6 +21,8 @@ export class PersonasComponent implements OnInit {
     this.getListaCrud()
   }
 
+  constructor(private toastr: ToastrService) { }
+
   getListaCrud() {
     this.loading = true
 
@@ -31,10 +34,15 @@ export class PersonasComponent implements OnInit {
   }
 
   borrarPersona(id: number) {
-    console.log(id);
-    this.personaServicio.deleteCrud(id).subscribe(()=> {
-      this.getListaCrud()
-    })
-  }
+    if (confirm(`Estas seguro que quieres borrar este usuario?`)) {
+      this.loading = true;
+      this.personaServicio.deleteCrud(id).subscribe(() => {
+        this.getListaCrud();
+        this.toastr.warning('La persona fué eliminado con exito', 'Persona eliminada')
+      })
+    } else {
+      this.toastr.info('La persona no fue eliminada', 'Persona no eliminada')
+    }
 
+  }
 }
